Code refactoring for Quick Fix
