+++ src/main/kotlin/com/github/dinbtechit/ngxs/action/editor/codeIntellisense/intention/NgxsActionIntentionAction.kt
@@ -1,7 +1,8 @@
 package com.github.dinbtechit.ngxs.action.editor.codeIntellisense.intention
 
-import com.github.dinbtechit.ngxs.action.editor.NgxsActionType
-import com.github.dinbtechit.ngxs.action.editor.NgxsStatePsiFile
+import com.github.dinbtechit.ngxs.action.editor.psi.state.NgxsActionType
+import com.github.dinbtechit.ngxs.action.editor.psi.state.NgxsStatePsiFileFactory
+import com.github.dinbtechit.ngxs.action.editor.psi.state.NgxsStatePsiUtil
 import com.intellij.codeInsight.intention.impl.BaseIntentionAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
@@ -21,12 +22,12 @@ class NgxsActionIntentionAction : BaseIntentionAction() {
 
     override fun isAvailable(project: Project, editor: Editor?, file: PsiFile?): Boolean {
         if (file?.virtualFile == null || editor == null) return false
-        return  NgxsStatePsiFile(file.virtualFile, project).isCursorWithinStateClass(editor, file)
+        return  NgxsStatePsiUtil.isCursorWithinStateClass(editor, file)
     }
 
     override fun invoke(project: Project, editor: Editor?, file: PsiFile?) {
         if (file?.virtualFile == null || editor == null) return
-        NgxsStatePsiFile(file.virtualFile, project).createActionMethodLiveTemplates(
+        NgxsStatePsiFileFactory.createActionMethodLiveTemplates(
             editor, file, NgxsActionType.WITHOUT_PAYLOAD
         )
     }
